requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
alembic=1.9.4=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.3=pypi_0
apptools=5.1.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
ase=3.22.1=pypi_0
asteval=0.9.28=pypi_0
astropy=4.3.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
bayesian-optimization=1.4.3=pypi_0
beautifulsoup4=4.11.1=pypi_0
bleach=5.0.1=pypi_0
bokeh=2.4.2=pypi_0
ca-certificates=2021.10.26=hecd8cb5_2
cached-property=1.5.2=pypi_0
cachetools=4.2.4=pypi_0
cairo=1.14.12=hc4e6be7_4
casm-python=1.2.1=pypi_0
certifi=2021.10.8=py37hecd8cb5_2
cffi=1.15.1=pypi_0
charset-normalizer=2.0.10=pypi_0
chemview=0.6=pypi_0
class-resolver=0.4.0=pypi_0
click=8.1.3=pypi_0
click-default-group=1.2.2=pypi_0
cma=3.3.0=pypi_0
cmaes=0.9.1=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.7.0=pypi_0
configobj=5.0.6=pypi_0
coverage=7.2.7=pypi_0
cycler=0.11.0=pypi_0
dataclasses-json=0.5.7=pypi_0
deap=1.3.3=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
django=2.2.26=pypi_0
django-crispy-forms=1.13.0=pypi_0
django-extensions=3.1.5=pypi_0
django-filter=1.1.0=pypi_0
djangorestframework=3.13.1=pypi_0
djangorestframework-filters=0.11.1=pypi_0
djangorestframework-queryfields=1.0.0=pypi_0
djangorestframework-xml=2.0.0=pypi_0
djangorestframework-yaml=2.0.0=pypi_0
dnspython=2.3.0=pypi_0
docdata=0.0.3=pypi_0
entrypoints=0.4=pypi_0
envisage=6.0.1=pypi_0
expat=2.4.1=h23ab428_2
fastjsonschema=2.16.1=pypi_0
ffmpeg=1.4=pypi_0
ffmpeg-python=0.2.0=pypi_0
flatbuffers=2.0=pypi_0
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.28.5=pypi_0
freetype=2.11.0=hd8bbffd_0
fribidi=1.0.10=haf1e3a3_0
frozenlist=1.3.0=pypi_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
gettext=0.21.0=h7535e17_0
glib=2.63.1=hd977a24_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=h38d11af_0
graphviz=2.40.1=hefbbd9a_2
greenlet=2.0.2=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
harfbuzz=1.8.8=hb8d4a28_0
hiphive=1.1=pypi_0
icu=58.2=h0a44026_3
idna=3.3=pypi_0
importlib-metadata=4.10.1=pypi_0
importlib-resources=5.7.0=pypi_0
inflect=6.0.5=pypi_0
iniconfig=1.1.1=pypi_0
ipykernel=6.15.2=pypi_0
ipymol=0.5=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h9ed2024_0
jsonschema=4.15.0=pypi_0
jupyter-client=7.3.5=pypi_0
jupyter-core=4.11.1=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.3=pypi_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
lark-parser=0.12.0=pypi_0
latexcodec=2.0.1=pypi_0
libclang=12.0.0=pypi_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.2.1=h0a44026_1007
libiconv=1.16=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.2.0=h87d7836_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=12.0.0=h0dcd299_1
llvmlite=0.39.1=pypi_0
lmfit=1.1.0=pypi_0
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h23ab428_1
mako=1.2.4=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matminer=0.8.0=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mayavi=4.7.4=pypi_0
metric-learn=0.6.2=pypi_0
mistune=2.0.4=pypi_0
mixture-of-experts=0.2.1=pypi_0
mock=5.0.1=pypi_0
monty=2022.1.19=pypi_0
more-click=0.1.2=pypi_0
more-itertools=9.0.0=pypi_0
mpmath=1.2.1=pypi_0
multidict=6.0.2=pypi_0
mypy-extensions=1.0.0=pypi_0
mysqlclient=2.1.0=pypi_0
nbclient=0.6.7=pypi_0
nbconvert=7.0.0=pypi_0
nbformat=5.4.0=pypi_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.5=pypi_0
networkx=2.6.3=pypi_0
nevergrad=0.8.0=pypi_0
notebook=6.4.12=pypi_0
numba=0.56.2=pypi_0
numpy=1.21.6=pypi_0
oauthlib=3.1.1=pypi_0
openssl=1.1.1m=hca72f7f_0
opt-einsum=3.3.0=pypi_0
optuna=3.1.0=pypi_0
packaging=21.3=pypi_0
palettable=3.3.0=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
pango=1.42.4=h7e27002_1
parso=0.8.3=pypi_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pypi_0
pickle5=0.0.12=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.0=pypi_0
pip=21.2.2=py37hecd8cb5_0
pixman=0.40.0=h9ed2024_1
pkgutil-resolve-name=1.3.10=pypi_0
plotly=5.5.0=pypi_0
pluggy=1.0.0=pypi_0
pooch=1.6.0=pypi_0
prisms-jobs=4.0.3=pypi_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.19.3=pypi_0
psutil=5.9.2=pypi_0
ptyprocess=0.7.0=pypi_0
pubchempy=1.0.4=pypi_0
pulp=2.6.0=pypi_0
py=1.11.0=pypi_0
py3dmol=1.8.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybtex=0.24.0=pypi_0
pycalphad=0.10.3=pypi_0
pycifrw=4.4.3=pypi_0
pycoranking=0.1.6=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.12=pypi_0
pydoe2=1.3.0=pypi_0
pydot=1.4.2=pypi_0
pyerfa=2.0.0.1=pypi_0
pyface=7.4.1=pypi_0
pygments=2.11.2=pypi_0
pygraphviz=1.3=py37h1de35cc_1
pykeen=1.8.1=pypi_0
pymatgen=2022.0.17=pypi_0
pymongo=4.4.1=pypi_0
pymysql=1.0.2=pypi_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
pyshtools=4.10=pypi_0
pystow=0.5.0=pypi_0
pytest=6.2.5=pypi_0
pytest-cov=4.1.0=pypi_0
python=3.7.7=hfe9666f_0_cpython
python-dateutil=2.8.2=pypi_0
python-graphviz=0.20.1=pypi_0
python-memcached=1.59=pypi_0
pytz=2021.3=pypi_0
pyxtal=0.5.2=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.1=pypi_0
qmpy=1.4.0=pypi_0
readline=7.0=h1de35cc_5
requests=2.27.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rexmex=0.1.3=pypi_0
robocrys=0.2.7=pypi_0
rsa=4.8=pypi_0
ruamel-yaml=0.17.20=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py37hecd8cb5_0
setuptools-scm=7.1.0=pypi_0
sh=1.14.3=pypi_0
shapely=1.8.0=pypi_0
six=1.16.0=pypi_0
sklearn=0.0=pypi_0
smt=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spglib=1.16.3=pypi_0
sqlalchemy=2.0.4=pypi_0
sqlite=3.33.0=hffcf06c_0
sqlparse=0.4.2=pypi_0
symengine=0.9.2=pypi_0
sympy=1.9=pypi_0
tabulate=0.8.9=pypi_0
tenacity=8.0.1=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.7.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.15.0=pypi_0
threadpoolctl=3.0.0=pypi_0
timeout-decorator=0.5.0=pypi_0
tinycss2=1.1.1=pypi_0
tinydb=4.8.0=pypi_0
tk=8.6.11=h7bc2e8c_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
torch=1.12.1=pypi_0
torch-geometric=2.3.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
trainstation=1.0=pypi_0
traitlets=5.3.0=pypi_0
traits=6.3.2=pypi_0
traitsui=7.3.1=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.8.0=pypi_0
umap=0.1.1=pypi_0
uncertainties=3.1.6=pypi_0
urllib3=1.26.8=pypi_0
vapory=0.1.2=pypi_0
vtk=9.1.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pypi_0
wrapt=1.13.3=pypi_0
wslink=1.5.3=pypi_0
xarray=0.20.2=pypi_0
xz=5.2.5=h1de35cc_0
yarl=1.7.2=pypi_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
